requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
supabase>=2.4.0
groq==0.4.2
rank-bm25==0.2.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
